Cleaned up code - strictly non-functional changes
