+++ src/Menus/ColorMenu.jsx
@@ -5,8 +5,7 @@ import { ColorPicker, ColorService } from "react-color-palette";
 // import "react-color-palette/css";
 // import "react-color-palette/dist/css/rcp.css";
 import options from "../options";
-import { Number } from "./MenuUtils";
-import { useAlignWithElement } from "./MenuHooks";
+import Number from "./Number";
 import Button from '@mui/material/Button';
 import ButtonGroup from '@mui/material/ButtonGroup';
 import Switch from '@mui/material/Switch';
