+++ src/index.jsx
@@ -2,5 +2,4 @@ import ReactDOM from 'react-dom/client';
 import './styling/index.css';
 import App from './App.jsx';
 
-
 ReactDOM.createRoot(document.getElementById('root')).render(<App />);
